fixing corner bug, improving camera dist support, cleaning

+++ physics/utils/RadixSort.hpp
@@ -47,7 +47,7 @@ class RadixSort
   bool createBuffers() const;
   bool createKernels() const;
 
-  unsigned int m_numEntities;
+  size_t m_numEntities;
 
   unsigned int m_numRadix;
